m5_uncertainty/models/quantile_lgbm/requirements.txt
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==5.4.1
scikit-learn==1.2.2
lightgbm==3.3.2
MAPIE==0.6.1